--- src/Interfaces/availabilityPanel.types.ts
@@ -1,10 +1,0 @@
-export type dragDropProps = {
-    idleMembers: Array<string>,
-    unAssignedTasks: Array<object | any>,
-  };
-
-export type droppableComponent = {
-    droppableId: string,
-    idleMembers: Array<string>,
-    unAssignedTasks: Array<object | any>,
-  };
