+++ src/components/Layout/index.tsx
@@ -21,34 +21,49 @@ const navBarContent = (title: string, refUrl: string) => {
   );
 };
 
-const Layout: FC<Props> = ({ children }) => (
-  <div className={styles.layout}>
-    <div className={styles.rdsHeader}>
-      <a href="https://realdevsquad.com" target="_blank" rel="noreferrer">
-        <Image width="45px" height="45px" className={styles.logo} src="/RDSLogo.png" alt="real-dev squad" />
-      </a>
-      <Link href="/">
-        <div className={styles.heading}>Status</div>
-      </Link>
+const Layout: FC<Props> = ({ children }) => {
+  const router = useRouter();
+
+  // Dev feature toggle
+  const { query } = router;
+  const dev = !!query.dev;
+
+  return (
+    <div className={styles.layout}>
+      <div className={styles.rdsHeader}>
+        <a href="https://realdevsquad.com" target="_blank" rel="noreferrer">
+          <Image width="45px" height="45px" className={styles.logo} src="/RDSLogo.png" alt="real-dev squad" />
+        </a>
+        <Link href="/">
+          <div className={styles.heading}>Status</div>
+        </Link>
+      </div>
+      <div className={styles.header}>
+        {navBarContent('Tasks', '/')}
+        |
+        {navBarContent('Mine', '/mine')}
+        |
+        {navBarContent('DS', '/challenges')}
+        |
+        {navBarContent('Open PRs', '/openPRs')}
+        |
+        {navBarContent('Stale PRs', '/stale-pr')}
+        |
+        {navBarContent('Idle Members', '/idle-members')}
+        {
+          (dev)
+            && (
+              <>
+                |
+                {navBarContent('Availability Panel', '/availability-panel')}
+              </>
+            )
+        }
+      </div>
+      {children}
+      <Footer />
     </div>
-    <div className={styles.header}>
-      {navBarContent('Tasks', '/')}
-      |
-      {navBarContent('Mine', '/mine')}
-      |
-      {navBarContent('DS', '/challenges')}
-      |
-      {navBarContent('Open PRs', '/openPRs')}
-      |
-      {navBarContent('Stale PRs', '/stale-pr')}
-      |
-      {navBarContent('Idle Members', '/idle-members')}
-      |
-      {navBarContent('Availability Panel', '/availability-panel')}
-    </div>
-    {children}
-    <Footer />
-  </div>
-);
+  );
+};
 
 export default Layout;
