+++ src/interfaces/availabilityPanel.type.ts
@@ -1,0 +1,12 @@
+import task from '@/interfaces/task.type';
+
+export type dragDropProps = {
+  idleMembers: Array<string>,
+  unAssignedTasks: Array<task>,
+};
+
+export type droppableComponent = {
+  droppableId: string,
+  idleMembers: Array<string>,
+  unAssignedTasks: Array<task>,
+};
