+++ src/interfaces/award.type.ts
@@ -1,0 +1,6 @@
+type award = {
+  dinero: number,
+  neelam: number,
+};
+
+export default award;
