+++ src/pages/availability-panel.tsx
@@ -2,7 +2,7 @@ import { FC, useState, useEffect } from 'react';
 import Head from '@/components/head';
 import Layout from '@/components/Layout';
 import useFetch from '@/hooks/useFetch';
-import { task } from '@/components/constants/types';
+import task from '@/interfaces/task.type';
 import classNames from '@/styles/availabilityPanel.module.scss';
 import DragDropcontext from '../components/availability-panel/drag-drop-context/index';
 
@@ -10,9 +10,9 @@ const TASKS_URL = `${process.env.NEXT_PUBLIC_BASE_URL}/tasks`;
 const IDLE_MEMBERS_URL = `${process.env.NEXT_PUBLIC_BASE_URL}/members/idle`;
 
 const AvailabilityPanel: FC = () => {
-  const [idleMembersList, setIdleMembersList] = useState<any>([]);
+  const [idleMembersList, setIdleMembersList] = useState<string[]>([]);
   const [tasks, setTasks] = useState<task[]>([]);
-  const [unAssignedTasks, setUnAssignedTasks] = useState<any>(null);
+  const [unAssignedTasks, setUnAssignedTasks] = useState<task[]>([]);
 
   const {
     response: taskResponse,
